requirements.txt
annotated-types==0.6.0
cachetools==5.3.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
data-traps-core @ git+ssh://git@github.com/mmazurekgda/data-traps-core.git@9014e76c1352c37d7f435c1fb5899a3bfbd126da
distlib==0.3.8
docstring-parser==0.15
filelock==3.13.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
identify==2.5.33
idna==3.6
importlib-metadata==7.0.0
iniconfig==2.0.0
kfp==2.0.0b13
kfp-pipeline-spec==0.2.0
kfp-server-api==2.0.0a6
kubernetes==23.6.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-yaml==1.2.0
pydantic_core==2.14.5
pytest==7.4.3
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
setuptools==68.2.2
six==1.16.0
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
websocket-client==1.7.0
wheel==0.41.2
zipp==3.17.0
